requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.1
SQLAlchemy==2.0.20
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.3.4
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.23.7
psycopg2-binary==2.9.9
pandas==2.2.3
tqdm==4.66.5
PyJWT>=2.8.0
